requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.1.5
idna==3.10
jira==3.8.0
lxml==5.3.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
python-docx==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sqlparse==0.5.3
TestLink-API-Python-client==0.8.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
